server/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-filter==24.3
python-decouple==3.8
psycopg2-binary==2.9.9